% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apify.R
\name{apify}
\alias{apify}
\title{Build an API URL}
\usage{
apify(
  fields,
  table_name,
  key,
  base_url = "https://api.civicore.com/newsos/api/v3/data/",
  page = 0
)
}
\arguments{
\item{fields}{Column names of data to be pulled from API connection.}

\item{table_name}{Name of table to draw from API.}

\item{key}{String providing log-in access to API.}

\item{base_url}{Home page of API.}

\item{page}{Default of 0, number of total pages in table is number records/100.}
}
\value{
A character vector
}
\description{
Build an API URL
}
\examples{
apify("fields","table_name","key","base_url")
}
