% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracketize.R
\name{bracketize}
\alias{bracketize}
\title{Convert character vector for API call}
\usage{
bracketize(x)
}
\arguments{
\item{x}{a character vector (or list) of field names}
}
\value{
a character vector with brackets and slahses
}
\description{
Convert character vector for API call
}
\examples{
bracketize(letters)
}
