% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{demographics}
\alias{demographics}
\title{Vector of participant ID fields}
\usage{
demographics()
}
\value{
a character vector of field names useful for API calls
}
\description{
Vector of participant ID fields
}
\examples{
demographics()
}
