% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_fields.R
\name{table_fields}
\alias{table_fields}
\title{Load list of field tables and names to be pulled from SOS's API}
\usage{
table_fields()
}
\value{
list of character vectors
}
\description{
Load list of field tables and names to be pulled from SOS's API
}
\examples{
table_fields()[1]
}
