% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sosRus-package.R
\docType{package}
\name{sosRus-package}
\alias{sosRus}
\alias{sosRus-package}
\title{sosRus: Tools to Assist Youth-Serving Organizations with Data Wrangling}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides a set of functions designed to assist youth-focused organizations in wrangling programmatic and participant-level data. Collectively, the functions in this package provide analysts with tools to improve research and evaluation practices by streamlining certain parts of the raw-to-analysis data pipeline.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ztimpe/sosRus}
  \item \url{https://ztimpe.github.io/sosRus/}
  \item Report bugs at \url{https://github.com/ztimpe/sosRus/issues}
}

}
\author{
\strong{Maintainer}: Zach Timpe \email{zacharytimpe@gmail.com}

}
\keyword{internal}
